w3c-reports/12092014/mac-wk2/runned-tests/IndexedDB/transaction-lifetime-expected.txt


FAIL Test events opening a second database when one connection is open already assert_equals: new version expected (number) 4 but got (object) null(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:330:19
http://localhost:8800/IndexedDB/transaction-lifetime.htm:36:30
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
